% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirstfunction.R
\name{myfirstfunction}
\alias{myfirstfunction}
\title{Title}
\usage{
myfirstfunction(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list with components 'x' and 'y', where 'y' is the square of 'x'
}
\description{
Title
}
\examples{
myfirstfunction(1:10)
}
