% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatterhist}
\alias{scatterhist}
\title{TA scatter plot with marginal histograms}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{a quantatative vector}

\item{y}{a quantatative vector}

\item{xlab}{x label}

\item{ylab}{y label}
}
\value{
a scatter plot of x vs y with marginal histograms
}
\description{
TA scatter plot with marginal histograms
}
\examples{
scatterhist(x=1:10, y=11:20)
}
